+++ packages/cli/src/commands/types.ts
@@ -1,1 +1,5 @@
 export type PackageManager = 'npm' | 'yarn' | 'pnpm'
+
+export type Structure = 'standalone' | 'monorepo'
+
+export type Template = 'js' | 'vue2' | 'vue3' | 'react'
